# openapi.yaml
openapi: 3.2.0
info:
  title: KantanDB HTTP API
  version: 0.1.0
  description: |
    A JSON document database API.

    Paths without a trailing slash are canonical. Requests with a trailing slash
    are redirected to the canonical path with 301 for GET and 307 for other
    methods.
servers:
  - url: /
...
paths:
...
  /db:
    get:
      tags: [Databases]
      operationId: listDatabases
      summary: List databases
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/DatabaseCursor"
      responses:
        "200":
          description: Database names in ascending order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseList"
              example:
                databases: [accounts, users]
        "400":
          $ref: "#/components/responses/BadRequest"
          description: The limit or cursor is invalid.
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    post:
      tags: [Databases]
      operationId: createDatabase
      summary: Create a database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Database"
            examples:
              withoutIndexes:
                value:
                  name: users
              withIndexes:
                value:
                  name: users
                  indexes:
                    - name: email
                      path: /email
      responses:
        "201":
          description: The database was created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Database"
        "400":
          $ref: "#/components/responses/BadRequest"
          description: The request, database name, or index definitions are invalid.
        "409":
          $ref: "#/components/responses/DatabaseExists"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /db/{database}:
    parameters:
      - $ref: "#/components/parameters/Database"

    get:
      tags: [Documents]
      operationId: listDocuments
      summary: List or query document IDs
      description: |
        Without `index` and `value`, lists every document ID. With both fields,
        queries one secondary index. `index` and `value` must appear together,
        and `op` may only appear with them. Unknown or repeated parameters are
        rejected. Equality accepts any JSON scalar. Ordering accepts only JSON
        numbers and strings. `value` is a URL-encoded JSON scalar, not a plain
        string.
      externalDocs:
        description: Query language
        url: docs/query-language.md
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/DocumentCursor"
        - $ref: "#/components/parameters/Index"
        - $ref: "#/components/parameters/Operator"
        - $ref: "#/components/parameters/IndexValue"
      responses:
        "200":
          description: A page of document IDs.
          headers:
            Accept-Query:
              $ref: "#/components/headers/AcceptQuery"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
        "400":
          $ref: "#/components/responses/BadRequest"
          description: The database name, query, limit, or cursor is invalid.
        "404":
          $ref: "#/components/responses/NotFound"
          description: The database or requested index does not exist.
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    post:
      tags: [Documents]
      operationId: createDocument
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Document"
            example:
              name: Alice
              active: true
      responses:
        "201":
          description: The document was created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
          description: The database name or document is invalid.
        "404":
          $ref: "#/components/responses/NotFound"
          description: The database does not exist.
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    query:
      tags: [Documents]
      operationId: queryDocuments
      summary: Query document IDs with JSONPath
      description: |
        Selects documents with an RFC 9535 JSONPath supplied in the request
        body. URI query parameters, unknown or duplicate fields, and trailing
        JSON are rejected. Equality accepts any JSON scalar. Ordering accepts
        only JSON numbers and strings. Requests are limited to 64 KiB and five
        seconds. Scan queries examine at most 10,000 documents per page.
      externalDocs:
        description: Query language
        url: docs/query-language.md
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathQuery"
            example:
              path: $.profile.age
              op: ge
              value: 18
              limit: 100
      responses:
        "200":
          description: A page of matching document IDs.
          headers:
            Accept-Query:
              $ref: "#/components/headers/AcceptQuery"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
        "400":
          $ref: "#/components/responses/BadRequest"
          description: The database name, request body, query, or cursor is invalid.
        "404":
          $ref: "#/components/responses/NotFound"
          description: The database does not exist.
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
          description: The service is unavailable or a query limit was exceeded.

    options:
      tags: [Documents]
      operationId: getDocumentCollectionOptions
      summary: Get document collection capabilities
      responses:
        "204":
          description: Collection capabilities.
          headers:
            Accept-Query:
              $ref: "#/components/headers/AcceptQuery"
            Allow:
              $ref: "#/components/headers/Allow"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    delete:
      tags: [Databases]
      operationId: deleteDatabase
      summary: Delete a database
      responses:
        "204":
          description: The database was deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
          description: The database name is invalid.
        "404":
          $ref: "#/components/responses/NotFound"
          description: The database does not exist.
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  schemas:
...
    Database:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          $ref: "#/components/schemas/Name"
        indexes:
          type: array
          maxItems: 16
          items:
            $ref: "#/components/schemas/Index"
    DatabaseList:
      type: object
      additionalProperties: false
      required: [databases]
      properties:
        databases:
          type: array
          items:
            $ref: "#/components/schemas/Name"
...
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results.
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    DatabaseCursor:
      name: cursor
      in: query
      description: Database name returned as the previous page position.
      schema:
        $ref: "#/components/schemas/Name"
...
  headers:
...
    Location:
      description: Relative path of the created resource.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalidRequest:
              value:
                error:
                  code: invalid_request
                  message: Could not read request body
            invalidName:
              value:
                error:
                  code: invalid_name
                  message: Database name is invalid
            invalidID:
              value:
                error:
                  code: invalid_id
                  message: Document ID is invalid
            invalidQuery:
              value:
                error:
                  code: invalid_query
                  message: Query is invalid
            invalidCursor:
              value:
                error:
                  code: invalid_cursor
                  message: Cursor is invalid
            invalidDocument:
              value:
                error:
                  code: invalid_document
                  message: Document must be a JSON object
            invalidPatch:
              value:
                error:
                  code: invalid_patch
                  message: Patch is invalid or produces a non-object document
            invalidIfMatch:
              value:
                error:
                  code: invalid_if_match
                  message: If-Match is invalid
...
    InternalError:
      description: Stored data is corrupt or an unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            internal:
              value:
                error:
                  code: internal_error
                  message: Internal server error
            corruption:
              value:
                error:
                  code: corrupt_data
                  message: Stored data is corrupt
    ServiceUnavailable:
      description: The service is unavailable or query execution exceeded a limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            unavailable:
              value:
                error:
                  code: service_unavailable
                  message: Service is unavailable
            timeout:
              value:
                error:
                  code: query_timeout
                  message: Query exceeded the execution limit
            queryLimit:
              value:
                error:
                  code: query_limit
                  message: Query exceeded the evaluation limit